Move code to completion.go
